% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LOLA.R
\name{buildRestrictedUniverse}
\alias{buildRestrictedUniverse}
\title{If you want to test for differential enrichment within
your usersets, you can restrict the universe to only
regions that are covered in at least one of your sets.
This function helps you build just such a restricted
universe}
\usage{
buildRestrictedUniverse(userSets)
}
\arguments{
\item{userSets}{The userSets you will pass to the enrichment calculation.}
}
\value{
A restricted universe
}
\description{
If you want to test for differential enrichment within
your usersets, you can restrict the universe to only
regions that are covered in at least one of your sets.
This function helps you build just such a restricted
universe
}
\examples{
data("sample_input", package="LOLA") # load userSets
restrictedUniverse = buildRestrictedUniverse(userSets)
}

