% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calcLocEnrichment.R
\name{calcLocEnrichment}
\alias{calcLocEnrichment}
\title{Workhorse function that calculates overlaps between userSets,
and then uses a fisher's exact test rank them by significance
of the overlap.}
\usage{
calcLocEnrichment(userSets, userUniverse, regionDB, dbTitle = "dbTitle",
  cores = 1, redefineUserSets = FALSE)
}
\arguments{
\item{userSets}{Regions of interest}

\item{userUniverse}{Regions tested for inclusion in userSets}

\item{regionDB}{Region DB to check for overlap, from loadRegionDB()}

\item{dbTitle}{deprecated parameter}

\item{cores}{Number of processors}

\item{redefineUserSets}{run redefineUserSets() on your userSets?}
}
\description{
Workhorse function that calculates overlaps between userSets,
and then uses a fisher's exact test rank them by significance
of the overlap.
}
\examples{
dbPath = system.file("extdata", "hg19", package="LOLA")
regionDB = loadRegionDB(dbLocation=dbPath)
data("sample_universe", package="LOLA")
data("sample_input", package="LOLA")
res = calcLocEnrichment(userSet, userUniverse, regionDB, dbTitle="dbTitle", cores=1)
locResult = res[2,]
extractEnrichmentOverlaps(locResult, userSet, regionDB)
}

