% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.R
\name{cleanws}
\alias{cleanws}
\title{cleanws takes multi-line, code formatted strings and just formats them
as simple strings}
\usage{
cleanws(string)
}
\arguments{
\item{string}{string to clean}
}
\value{
A string with all consecutive whitespace characters, including
tabs and newlines, merged into a single space.
}
\description{
cleanws takes multi-line, code formatted strings and just formats them
as simple strings
}

