% Generated by roxygen2 (4.0.1): do not edit by hand
\name{countOverlapsAnyRev}
\alias{countOverlapsAnyRev}
\title{Just a reverser. Reverses the order of arguments and passes them untouched to countOverlapsAny -- so you can use it with lapply.}
\usage{
countOverlapsAnyRev(subj, quer)
}
\description{
Just a reverser. Reverses the order of arguments and passes them untouched to countOverlapsAny -- so you can use it with lapply.
}

