% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extractEnrichmentOverlaps}
\alias{extractEnrichmentOverlaps}
\title{this function, given a single row from an enrichment table calculation, it will find the set of overlaps between the user set and the test set. You can then use these, for example, to get fasta sequences for those regions.}
\usage{
extractEnrichmentOverlaps(locResult, userSets)
}
\description{
this function, given a single row from an enrichment table calculation, it will find the set of overlaps between the user set and the test set. You can then use these, for example, to get fasta sequences for those regions.
}

