% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getRegionGroupSizes}
\alias{getRegionGroupSizes}
\title{Given a region annotation object, this function either reads its sizes
file if one exists, or otherwise calculates sizes and creates a sizes
file.}
\usage{
getRegionGroupSizes(DT, bedDir, shareDir = getOption("SHARE.DATA.DIR"))
}
\arguments{
\item{dbLocation}{folder of regiondB}

\item{annoDT}{the result of readRegionSetAnnotation();}
}
\description{
Given a region annotation object, this function either reads its sizes
file if one exists, or otherwise calculates sizes and creates a sizes
file.
}

