% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loadRegionDB.R
\name{getRegionSet}
\alias{getRegionSet}
\title{Grab a single region set from a database, specified by filename.}
\usage{
getRegionSet(regionDB, filenames, collections = NULL)
}
\arguments{
\item{regionDB}{A region database loaded with loadRegionDB().}

\item{fn}{Filename of a particular region set to grab.}
}
\value{
A GRanges object derived from the specified file in the regionDB.
}
\description{
If you want to work with a LOLA regionDB region set individually,
this function can help you. It can extract individual (or subsets of)
region sets from either loaded regionDBs, loaded with loadRegionDB(), or
from a database on disk, where only the region sets of interest will
be loaded.
}
\examples{
examples/example.R
}

