% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loadRegionDB.R
\name{listRegionSets}
\alias{listRegionSets}
\title{Lists the region sets for given collection(s) in a region database on disk.}
\usage{
listRegionSets(regionDB, collections = NULL)
}
\arguments{
\item{regionDB}{File path to region database}

\item{collections}{(optional) subset of collections to list}
}
\value{
a list of files in the given collections
}
\description{
Lists the region sets for given collection(s) in a region database on disk.
}
\examples{
dbPath = system.file("extdata", "hg19", package="LOLA")
listRegionSets(dbPath)
}

