% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{listToGRangesList}
\alias{listToGRangesList}
\title{converts a list of GRanges into a GRangesList; strips all metadata.}
\usage{
listToGRangesList(lst)
}
\arguments{
\item{lst}{a list of GRanges objects}
}
\value{
a GRangesList object
}
\description{
converts a list of GRanges into a GRangesList; strips all metadata.
}

