% Generated by roxygen2 (4.0.1): do not edit by hand
\name{loadAllEnrichmentDatabases}
\alias{loadAllEnrichmentDatabases}
\title{Loads All Enrichment Databases.}
\usage{
loadAllEnrichmentDatabases()
}
\description{
Helper loader functions to just load up all the data, if you want
to do a comprehensive analysis.
}
\examples{
loadAllEnrichmentDatabases();
}

