% Generated by roxygen2 (4.0.1): do not edit by hand
\name{loadLocationEnrichmentDatabases}
\alias{loadLocationEnrichmentDatabases}
\title{Loads all location databases. Just a helper function that calls the others.
use loadLocationEnrichmentMm10() for mouse.}
\usage{
loadLocationEnrichmentDatabases()
}
\description{
Loads all location databases. Just a helper function that calls the others.
use loadLocationEnrichmentMm10() for mouse.
}

