% Generated by roxygen2 (4.0.1): do not edit by hand
\name{loadRegionDB}
\alias{loadRegionDB}
\title{Helper function to annotate and load a regionDB, a folder with
subfolder collections of regions.}
\usage{
loadRegionDB(dbLocation, filePattern = "", limit = NULL)
}
\arguments{
\item{dbLocation}{folder where your regionDB is stored.}

\item{filePattern}{passed to list.files; you can use this to select only certain file names in your folders.}
}
\description{
Helper function to annotate and load a regionDB, a folder with
subfolder collections of regions.
}
\examples{
regionDB = loadRegionDB(dbLocation= "~/fhgfs/share/regionDB/hg19")
}

