% Generated by roxygen2 (4.0.1): do not edit by hand
\name{locationEnrichment}
\alias{locationEnrichment}
\title{Convenience function for doing all the location enrichment functions
in one fell swoop.}
\usage{
locationEnrichment(userSets, userUniverse, checkUniverse = FALSE, cores = 4,
  redefineUserSets = TRUE)
}
\description{
Convenience function for doing all the location enrichment functions
in one fell swoop.
}

