% Generated by roxygen2 (4.0.1): do not edit by hand
\name{locationEnrichmentRegionDB}
\alias{locationEnrichmentRegionDB}
\title{Convenience function for doing all the location enrichment functions
in one fell swoop.
designed to use new generic scripts.}
\usage{
locationEnrichmentRegionDB(userSets, userUniverse, regionDB,
  checkUniverse = FALSE, cores = 4, redefineUserSets = TRUE)
}
\description{
Convenience function for doing all the location enrichment functions
in one fell swoop.
designed to use new generic scripts.
}

