% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{LOLA}
\alias{LOLA}
\alias{LOLA-package}
\title{Provides functions for genome location overlap analysis. Further documentation forthcoming.}
\description{
LOLA!
}
\author{
Nathan Sheffield
}
\references{
\url{http://github.com/sheffield}
}

