% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadRegionDB.R
\name{mergeRegionDBs}
\alias{mergeRegionDBs}
\title{Given two regionDBs, (lists returned from loadRegionDB()),
This function will combine them into a single regionDB. This
will enable you to combine, for example, LOLA Core databases
with custom databases into a single analysis.}
\usage{
mergeRegionDBs(dbA, dbB)
}
\arguments{
\item{dbA}{First regionDB database.}

\item{dbB}{Second regionDB database.}
}
\value{
A combined regionDB.
}
\description{
Given two regionDBs, (lists returned from loadRegionDB()),
This function will combine them into a single regionDB. This
will enable you to combine, for example, LOLA Core databases
with custom databases into a single analysis.
}
\examples{
dbPath = system.file("extdata", "hg19", package="LOLA")
regionDB = loadRegionDB(dbPath)
combinedRegionDB = mergeRegionDBs(regionDB, regionDB)
}

