% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{nlist}
\alias{nlist}
\title{Named list function.}
\usage{
nlist(...)
}
\arguments{
\item{...}{arguments passed to list()}
}
\value{
A named list object.
}
\description{
This function is a drop-in replacement for the base list() function,
which automatically names your list according to the names of the
variables used to construct it.
It seemlessly handles lists with some names and others absent,
not overwriting specified names while naming any unnamed parameters.
Took me awhile to figure this out.
}

