% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.R
\name{readBed}
\alias{readBed}
\title{Imports bed files and creates GRanges objects, using the fread()
function from data.table.}
\usage{
readBed(file)
}
\arguments{
\item{file}{File name of bed file.}
}
\description{
Imports bed files and creates GRanges objects, using the fread()
function from data.table.
}

