% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/loadRegionDB.R
\name{readCollectionAnnotation}
\alias{readCollectionAnnotation}
\title{Read collection annotation}
\usage{
readCollectionAnnotation(dbLocation)
}
\arguments{
\item{dbLocation}{Location of the database}
}
\description{
Read collection annotation
}
\examples{
dbPath = system.file("extdata", "hg19", package="LOLA")
collectionAnno = readCollectionAnnotation(dbLocation=dbPath)
}

