% Generated by roxygen2 (4.0.1): do not edit by hand
\name{readRegionAnnotation}
\alias{readRegionAnnotation}
\title{Given a folder containing region collections in subfolders, this function
will either read the annotation file if one exists, or create a generic
annotation file.}
\usage{
readRegionAnnotation(shareDir = getOption("SHARE.DATA.DIR"),
  bedDir = "regionDB/bock_regions_mm10/")
}
\description{
Given a folder containing region collections in subfolders, this function
will either read the annotation file if one exists, or create a generic
annotation file.
}

