% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/loadRegionDB.R
\name{readRegionGRL}
\alias{readRegionGRL}
\title{This function takes a region annotation object and reads in the regions,
returning a GRangesList object of the regions.}
\usage{
readRegionGRL(dbLocation, annoDT, useCache = TRUE, limit = NULL)
}
\arguments{
\item{dbLocation}{folder of regiondB}

\item{annoDT}{output of readRegionSetAnnotation().}

\item{useCache}{uses simpleCache to cache and load the results}

\item{limit}{for testing purposes, limit the number of files read. NULL for no limit (default).}
}
\description{
This function takes a region annotation object and reads in the regions,
returning a GRangesList object of the regions.
}

