% Generated by roxygen2 (4.0.1): do not edit by hand
\name{readRegionSetAnnotation}
\alias{readRegionSetAnnotation}
\title{Given a folder containing region collections in subfolders, this function
will either read the annotation file if one exists, or create a generic
annotation file.}
\usage{
readRegionSetAnnotation(dbLocation, filePattern = "", refreshSizes = FALSE)
}
\description{
Given a folder containing region collections in subfolders, this function
will either read the annotation file if one exists, or create a generic
annotation file.
}

