% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{replaceFileExtension}
\alias{replaceFileExtension}
\title{This will change the string in filename to have a new extension}
\usage{
replaceFileExtension(filename, extension)
}
\arguments{
\item{filename}{string to convert}

\item{extension}{new extension}
}
\value{
Filename with original extension deleted, replaced by provided
extension
}
\description{
This will change the string in filename to have a new extension
}

