% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.R
\name{setSharedDataDir}
\alias{setSharedDataDir}
\title{setSharedDataDir
Sets global variable specifying the default data directory.}
\usage{
setSharedDataDir(sharedDataDir)
}
\arguments{
\item{sharedDataDir}{directory where the shared data is stored.}
}
\description{
setSharedDataDir
Sets global variable specifying the default data directory.
}

