% Generated by roxygen2 (4.0.1): do not edit by hand
\name{splitFileIntoCollection}
\alias{splitFileIntoCollection}
\title{This function will take a single large bed file and split it into files
for use with the LOLA collection format.}
\usage{
splitFileIntoCollection(filename, splitCol)
}
\description{
This function will take a single large bed file and split it into files
for use with the LOLA collection format.
}

