% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{userSet}
\alias{userSet}
\title{An example set of regions, sampled from the example database.}
\format{A GRanges object}
\usage{
data(sample_input)
}
\description{
A dataset containing a few sample regions.
}
\keyword{datasets}

