% Generated by roxygen2 (4.0.1): do not edit by hand
\name{writeCombinedEnrichment}
\alias{writeCombinedEnrichment}
\title{Function for writing output all at once: combinedResults is an table
generated by "locationEnrichment()".# or by rbinding category and location results.
Writes all enrichments to a single file, and also spits out the same data
divided into groups based on userSets, and Databases, just for convenience.
disable this with an option.}
\usage{
writeCombinedEnrichment(combinedResults, outFolder = NULL,
  includeSplits = TRUE)
}
\description{
Function for writing output all at once: combinedResults is an table
generated by "locationEnrichment()".# or by rbinding category and location results.
Writes all enrichments to a single file, and also spits out the same data
divided into groups based on userSets, and Databases, just for convenience.
disable this with an option.
}

