% Generated by roxygen2 (4.0.1): do not edit by hand
\name{writeDataTableSplitByColumn}
\alias{writeDataTableSplitByColumn}
\title{Given a data table and a factor variable to split on,
efficiently divides the table and then writes the different splits
to separate files, named with filePrepend and numbered according
to split.}
\usage{
writeDataTableSplitByColumn(DT, splitFactor, filePrepend = "",
  orderColumn = NULL)
}
\value{
number of splits written
}
\description{
Given a data table and a factor variable to split on,
efficiently divides the table and then writes the different splits
to separate files, named with filePrepend and numbered according
to split.
}

