% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utility.R
\name{countOverlapsAnyRev}
\alias{countOverlapsAnyRev}
\title{Just a reverser. Reverses the order of arguments and passes them
untouched to countOverlapsAny -- so you can use it with lapply.}
\usage{
countOverlapsAnyRev(subj, quer)
}
\arguments{
\item{subj}{Subject}

\item{quer}{Query}
}
\value{
Results from countOverlaps
}
\description{
Just a reverser. Reverses the order of arguments and passes them
untouched to countOverlapsAny -- so you can use it with lapply.
}

