% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/postEnrichment.R
\name{extractEnrichmentOverlaps}
\alias{extractEnrichmentOverlaps}
\title{Given a single row from an enrichment table calculation,
finds the set of overlaps between the user set and the test set.
You can then use these, for example, to get sequences for those regions.}
\usage{
extractEnrichmentOverlaps(locResult, userSets, regionDB)
}
\arguments{
\item{locResult}{Results from enrichmentCalc function}

\item{userSets}{User sets passed to the enrichmentCalc function}

\item{regionDB}{Region database used}
}
\description{
Given a single row from an enrichment table calculation,
finds the set of overlaps between the user set and the test set.
You can then use these, for example, to get sequences for those regions.
}
\examples{
dbPath = system.file("extdata", "hg19", package="LOLA")
regionDB = loadRegionDB(dbLocation=dbPath)
data("sample_universe", package="LOLA")
data("sample_input", package="LOLA")
res = runLOLA(userSet, userUniverse, regionDB, cores=1)
locResult = res[2,]
extractEnrichmentOverlaps(locResult, userSet, regionDB)
}

