% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{lapplyAlias}
\alias{lapplyAlias}
\title{Function to run lapply or mclapply, depending on the option set in
getOption("mc.cores"), which can be set with setLapplyAlias().}
\usage{
lapplyAlias(..., mc.preschedule = TRUE)
}
\arguments{
\item{...}{Arguments passed lapply() or mclapply()}

\item{mc.preschedule}{Argument passed to mclapply}
}
\value{
Result from lapply or parallel::mclapply
}
\description{
Function to run lapply or mclapply, depending on the option set in
getOption("mc.cores"), which can be set with setLapplyAlias().
}

