% Generated by roxygen2 (4.0.1): do not edit by hand
\name{listToGRangesList}
\alias{listToGRangesList}
\title{converts a list of GRanges into a GRangesList; strips all metadata.}
\usage{
listToGRangesList(lst)
}
\arguments{
\item{lst}{a list of GRanges objects}
}
\value{
a GRangesList object
}
\description{
converts a list of GRanges into a GRangesList; strips all metadata.
}

