% Generated by roxygen2 (4.0.1): do not edit by hand
\name{loadRegionDB}
\alias{loadRegionDB}
\title{Helper function to annotate and load a regionDB, a folder with
subfolder collections of regions.}
\usage{
loadRegionDB(dbLocation, filePattern = "")
}
\description{
Helper function to annotate and load a regionDB, a folder with
subfolder collections of regions.
}
\examples{
regionDB = loadRegionDB(dbLocation= "~/fhgfs/share/regionDB/hg19")
}

