% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LOLA.R
\docType{package}
\name{LOLA}
\alias{LOLA}
\alias{LOLA-package}
\title{Genome locus overlap analysis.}
\description{
Run, Lola!
}
\author{
Nathan Sheffield
}
\references{
\url{http://github.com/sheffien}
}

