% Generated by roxygen2 (4.0.1): do not edit by hand
\name{readRegionGRL}
\alias{readRegionGRL}
\title{This function takes a region annotation object and reads in the regions,
returning a GRangesList object of the regions.}
\usage{
readRegionGRL(dbLocation, annoDT, limit = NULL)
}
\arguments{
\item{dbLocation}{folder of regiondB}

\item{annoDT}{output of readRegionAnnotation().}

\item{limit}{for testing purposes, you could limit the number of files read. NULL for no limit (default).}
}
\description{
This function takes a region annotation object and reads in the regions,
returning a GRangesList object of the regions.
}

