% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loadRegionDB.R
\name{readRegionSetAnnotation}
\alias{readRegionSetAnnotation}
\title{Given a folder containing region collections in subfolders, this function
will either read the annotation file if one exists, or create a generic
annotation file.}
\usage{
readRegionSetAnnotation(dbLocation, filePattern = "", refreshSizes = FALSE)
}
\arguments{
\item{dbLocation}{folder where your regionDB is stored.}

\item{filePattern}{passed to list.files; you can use this
to select only certain file names in your folders.}

\item{refreshSizes}{should I recreate the sizes files
    documenting how many regions (lines) are in each region set?}
}
\value{
Region set annotation (data.table)
}
\description{
Given a folder containing region collections in subfolders, this function
will either read the annotation file if one exists, or create a generic
annotation file.
}
\examples{
dbPath = system.file("extdata", "hg19", package="LOLA")
regionAnno = readRegionSetAnnotation(dbLocation=dbPath)
}

