% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.R
\name{sampleGRL}
\alias{sampleGRL}
\title{Function to sample regions from a GRangesList object, in specified proportion}
\usage{
sampleGRL(GRL, prop)
}
\arguments{
\item{GRL}{GRangesList from which to sample}

\item{prop}{vector with same length as GRL, of values between 0-1,
proportion of the list to select}
}
\value{
A sampled subset of original GRangesList object.
}
\description{
Function to sample regions from a GRangesList object, in specified proportion
}

