% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{setLapplyAlias}
\alias{setLapplyAlias}
\title{To make parallel processing a possibility but not required,
I use an lapply alias which can point at either the base lapply
(for no multicore), or it can point to mclapply,
and set the options for the number of cores (what mclapply uses).
With no argument given, returns intead the number of cpus currently selected.}
\usage{
setLapplyAlias(cores = 0)
}
\arguments{
\item{cores}{Number of cpus}
}
\value{
None
}
\description{
To make parallel processing a possibility but not required,
I use an lapply alias which can point at either the base lapply
(for no multicore), or it can point to mclapply,
and set the options for the number of cores (what mclapply uses).
With no argument given, returns intead the number of cpus currently selected.
}
