% Generated by roxygen2 (4.0.1): do not edit by hand
\name{setSharedDataDir}
\alias{setSharedDataDir}
\title{setSharedDataDir
Sets global variable specifying the default data directory.}
\usage{
setSharedDataDir(sharedDataDir)
}
\description{
setSharedDataDir
Sets global variable specifying the default data directory.
}

