% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postEnrichment.R
\name{splitDataTable}
\alias{splitDataTable}
\title{Efficiently split a data.table by a column in the table}
\usage{
splitDataTable(DT, splitFactor)
}
\arguments{
\item{DT}{Data.table to split}

\item{splitFactor}{Column to split, which can be a character vector
or an integer.}
}
\value{
List of data.table objects, split by column
}
\description{
Efficiently split a data.table by a column in the table
}

