% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/postEnrichment.R
\name{splitDataTable}
\alias{splitDataTable}
\title{Efficiently splits a data.table}
\usage{
splitDataTable(DT, splitFactor)
}
\arguments{
\item{DT}{data.table object to split}

\item{splitFactor}{column of DT to split on}
}
\description{
Efficiently splits a data.table
}

