% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{userUniverse}
\alias{userUniverse}
\title{A reduced GRanges object from the example regionDB database}
\format{A GRanges object}
\usage{
data(sample_universe)
}
\description{
A reduced GRanges object from the example regionDB database
}
\keyword{datasets}

