% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.R
\name{write.tsv}
\alias{write.tsv}
\title{Wrapper of write.table that provides defaults to write a
simple .tsv file. Passes additional arguments to write.table}
\usage{
write.tsv(...)
}
\arguments{
\item{...}{Additional arguments passed to write.table}
}
\description{
Wrapper of write.table that provides defaults to write a
simple .tsv file. Passes additional arguments to write.table
}

