% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/postEnrichment.R
\name{writeCombinedEnrichment}
\alias{writeCombinedEnrichment}
\title{Function for writing output all at once: combinedResults is an table
generated by "locationEnrichment()" or by rbinding category and location results.
Writes all enrichments to a single file, and also spits out the same data
divided into groups based on userSets, and Databases, just for convenience.
disable this with an option.}
\usage{
writeCombinedEnrichment(combinedResults, outFolder = NULL,
  includeSplits = TRUE)
}
\arguments{
\item{combinedResults}{enrichment results object}

\item{outFolder}{location to write results on disk}

\item{includeSplits}{also include individual files for each user set and database?}
}
\value{
No return value.
}
\description{
Function for writing output all at once: combinedResults is an table
generated by "locationEnrichment()" or by rbinding category and location results.
Writes all enrichments to a single file, and also spits out the same data
divided into groups based on userSets, and Databases, just for convenience.
disable this with an option.
}

