% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postEnrichment.R
\name{writeDataTableSplitByColumn}
\alias{writeDataTableSplitByColumn}
\title{Given a data table and a factor variable to split on,
efficiently divides the table and then writes the different splits
to separate files, named with filePrepend and numbered according
to split.}
\usage{
writeDataTableSplitByColumn(
  DT,
  splitFactor,
  filePrepend = "",
  orderColumn = NULL
)
}
\arguments{
\item{DT}{data.table to split}

\item{splitFactor}{column of DT to split on}

\item{filePrepend}{notation string to prepend to output files}

\item{orderColumn}{column of DT to order on (defaults to the first column)}
}
\value{
number of splits written
}
\description{
Given a data table and a factor variable to split on,
efficiently divides the table and then writes the different splits
to separate files, named with filePrepend and numbered according
to split.
}
